+++ R/add_p.R
@@ -45,8 +45,7 @@
 #' @author Emily C. Zabor, Daniel D. Sjoberg
 #' @examples
 #' add_p_ex1 <-
-#'   trial %>%
-#'   dplyr::select(age, grade, response, trt) %>%
+#'   trial[c("age", "grade", "response", "trt")] %>%
 #'   tbl_summary(by = trt) %>%
 #'   add_p()
 #'
@@ -65,7 +64,7 @@
 #' add_p_ex2 <-
 #'   trial[c("response", "trt")] %>%
 #'   tbl_summary(by = trt) %>%
-#'   add_p(test = vars(response) ~ "my_mcnemar")
+#'   add_p(test = response ~ "my_mcnemar")
 #' }
 #' @section Example Output:
 #' \if{html}{Example 1}
@@ -80,31 +79,18 @@ add_p <- function(x, test = NULL, pvalue_fun = NULL,
                   group = NULL, include = NULL, exclude = NULL) {
 
   # converting bare arguments to string ----------------------------------------
-  group <- enquo_to_string(rlang::enquo(group), arg_name = "group")
-
-  # putting arguments in a list to pass to tbl_summary_
-  add_p_args <- as.list(environment())
-
-  # passing arguments to add_p_
-  do.call(add_p_, add_p_args)
-}
-
-#' Standard evaluation version of add_p()
-#'
-#' The `'group ='` argument can be passed as a string, rather than with non-standard
-#' evaluation as in [add_p]. Review the help file for [add_p] fully documented
-#' options and arguments.
-#'
-#' @inheritParams add_p
-#' @export
-add_p_ <- function(x, test = NULL, pvalue_fun = NULL,
-                   group = NULL, include = NULL, exclude = NULL) {
+  group <- var_input_to_string(data = x$inputs$data, select_input = !!rlang::enquo(group),
+                               arg_name = "by", select_single = TRUE)
+  include <- var_input_to_string(data = x$inputs$data, select_input = !!rlang::enquo(include),
+                                 arg_name = "by")
+  exclude <- var_input_to_string(data = x$inputs$data, select_input = !!rlang::enquo(exclude),
+                                 arg_name = "by")
 
   # group argument -------------------------------------------------------------
   if (!is.null(group)) {
     # checking group is in the data frame
     if (!group %in% x$meta_data$variable) {
-      stop(glue("'{group}' is not a column name in the input data frame."))
+      stop(glue("'{group}' is not a column name in the input data frame."), call. = FALSE)
     }
     # dropping group variable from table_body and meta_data
     x$table_body <- x$table_body %>% filter(.data$variable != group)
@@ -120,41 +106,41 @@ add_p_ <- function(x, test = NULL, pvalue_fun = NULL,
       "'pvalue_fun' is not a valid function.  Please pass only a function\n",
       "object. For example,\n\n",
       "'pvalue_fun = function(x) style_pvalue(x, digits = 2)'"
-    ))
+    ), call. = FALSE)
   }
 
   # checking that input is class tbl_summary
-  if (class(x) != "tbl_summary") stop("x must be class 'tbl_summary'")
+  if (class(x) != "tbl_summary") stop("x must be class 'tbl_summary'", call. = FALSE)
   # checking that input x has a by var
   if (is.null(x$inputs[["by"]])) {
     stop(paste0(
       "Cannot add comparison when no 'by' variable ",
       "in original tbl_summary() call"
-    ))
+    ), call. = FALSE)
   }
 
   # test -----------------------------------------------------------------------
   # parsing into a named list
   test <- tidyselect_to_list(
     x$inputs$data, test,
-    .meta_data = x$meta_data, input_type = "test"
+    .meta_data = x$meta_data, arg_name = "test"
   )
 
   if (!is.null(test)) {
     # checking that all inputs are named
     if ((names(test) %>%
-      purrr::discard(. == "") %>%
-      length()) != length(test)) {
+         purrr::discard(. == "") %>%
+         length()) != length(test)) {
       stop(glue(
         "Each element in 'test' must be named. ",
         "For example, 'test = list(age = \"t.test\", ptstage = \"fisher.test\")'"
-      ))
+      ), call. = FALSE)
     }
   }
 
   # checking pvalue_fun are functions
   if (!is.function(pvalue_fun)) {
-    stop("Input 'pvalue_fun' must be a function.")
+    stop("Input 'pvalue_fun' must be a function.", call. = FALSE)
   }
 
   # Getting p-values only for included variables