+++ R/tbl_stack.R
@@ -19,7 +19,7 @@
 #'   glm(response ~ trt, trial, family = binomial) %>%
 #'   tbl_regression(
 #'     exponentiate = TRUE,
-#'     label = list(vars(trt) ~ "Treatment (unadjusted)")
+#'     label = list(trt ~ "Treatment (unadjusted)")
 #'   )
 #'
 #' t2 <-
@@ -27,7 +27,7 @@
 #'   tbl_regression(
 #'     include = "trt",
 #'     exponentiate = TRUE,
-#'     label = list(vars(trt) ~ "Treatment (adjusted)")
+#'     label = list(trt ~ "Treatment (adjusted)")
 #'   )
 #'
 #' tbl_stack_ex1 <- tbl_stack(list(t1, t2))
@@ -38,7 +38,7 @@
 #'   coxph(Surv(ttdeath, death) ~ trt, trial) %>%
 #'   tbl_regression(
 #'     exponentiate = TRUE,
-#'     label = list(vars(trt) ~ "Treatment (unadjusted)")
+#'     label = list(trt ~ "Treatment (unadjusted)")
 #'   )
 #'
 #' t4 <-
@@ -46,7 +46,7 @@
 #'   tbl_regression(
 #'     include = "trt",
 #'     exponentiate = TRUE,
-#'     label = list(vars(trt) ~ "Treatment (adjusted)")
+#'     label = list(trt ~ "Treatment (adjusted)")
 #'   )
 #'
 #'